refactor: remove duplicated tests
