+++ test/uglifyOptions-option.test.js
@@ -9,11 +9,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `ecma` option (value is `5`)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/ecma-5/entry.js`,
-      output: {
-        path: `${__dirname}/dist-ecma-5`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin({
@@ -45,11 +40,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `ecma` option (value is `6`)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/ecma-6/entry.js`,
-      output: {
-        path: `${__dirname}/dist-ecma-6`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin({
@@ -81,11 +71,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `ecma` option (value is `7`)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/ecma-7/entry.js`,
-      output: {
-        path: `${__dirname}/dist-ecma-7`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -116,11 +101,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `ecma` option (value is `8`)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/ecma-8/entry.js`,
-      output: {
-        path: `${__dirname}/dist-ecma-8`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin({
@@ -152,11 +132,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `warnings` option (boolean `true` value)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/unreachable-code.js`,
-      output: {
-        path: `${__dirname}/dist`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin({
@@ -183,11 +158,6 @@ describe('when applied with uglifyOptions options', () => {
   it('matches snapshot for `warnings` option (boolean `false` value)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/unreachable-code.js`,
-      output: {
-        path: `${__dirname}/dist`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin({
@@ -212,14 +182,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `parse` option (object value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -245,14 +208,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `compress` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -276,14 +232,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `compress` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -307,14 +256,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `compress` option (object value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -340,14 +282,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `mangle` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -371,14 +306,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `mangle` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -402,14 +330,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `mangle` option (object value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -434,15 +355,8 @@ describe('when applied with uglifyOptions options', () => {
     });
   });
 
-  it('matches snapshot for `output` option (object value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+  it('matches snapshot for `output` option (object value with `true` value for `beautify`)', () => {
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -467,15 +381,34 @@ describe('when applied with uglifyOptions options', () => {
     });
   });
 
-  it('matches snapshot for `toplevel` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
+  it('matches snapshot for `output` option (object value with `true` value for `comments`)', () => {
+    const compiler = createCompiler();
+
+    new UglifyJsPlugin({
+      uglifyOptions: {
+        output: {
+          comments: true,
+        },
       },
+    }).apply(compiler);
+
+    return compile(compiler).then((stats) => {
+      const errors = stats.compilation.errors.map(cleanErrorStack);
+      const warnings = stats.compilation.warnings.map(cleanErrorStack);
+
+      expect(errors).toMatchSnapshot('errors');
+      expect(warnings).toMatchSnapshot('warnings');
+
+      for (const file in stats.compilation.assets) {
+        if (Object.prototype.hasOwnProperty.call(stats.compilation.assets, file)) {
+          expect(stats.compilation.assets[file].source()).toMatchSnapshot(file);
+        }
+      }
     });
+  });
+
+  it('matches snapshot for `toplevel` option (boolean `true` value)', () => {
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -499,14 +432,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `toplevel` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -530,14 +456,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `nameCache` option (empty object value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -561,14 +480,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `ie8` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -592,14 +504,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `ie8` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -623,14 +528,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `keep_classnames` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -654,14 +552,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `keep_classnames` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -685,14 +576,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `keep_fnames` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -716,14 +600,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `keep_fnames` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -747,14 +624,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `safari10` option (boolean `true` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -778,14 +648,7 @@ describe('when applied with uglifyOptions options', () => {
   });
 
   it('matches snapshot for `safari10` option (boolean `false` value)', () => {
-    const compiler = createCompiler({
-      entry: `${__dirname}/fixtures/import-export/entry.js`,
-      output: {
-        path: `${__dirname}/dist-import-export`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
-    });
+    const compiler = createCompiler();
 
     new UglifyJsPlugin({
       uglifyOptions: {
@@ -811,11 +674,6 @@ describe('when applied with uglifyOptions options', () => {
   it('disable inline optimization by default (have a lot of problems)', () => {
     const compiler = createCompiler({
       entry: `${__dirname}/fixtures/inline-optimization.js`,
-      output: {
-        path: `${__dirname}/dist-inline-optimization`,
-        filename: '[name].js',
-        chunkFilename: '[id].[name].js',
-      },
     });
 
     new UglifyJsPlugin().apply(compiler);
